Fix tasks not being cached
The problem here is that if the `cargoManifestDir` is marked as task
input, the task can never be up to date, because every run of the task
will change something in its "target" subdirectory, instanstly invalidating
the cache (which is based on the state of the directory before start of
the task execution).

Instead, we now only use the manifest and the "src" directory as task
input. With this change, the task cache is only invalidated if something
actually changed. Note that we are not being 100% correct here, for that
we would have to take into account all `cargo:rerun-if-changed` build
directives. And the actual source directory of the lib could also be
different than "src". Getting this information is much much harder though
and not required for iroh-ffi specifically, so we kinda just ignore that.

This simple change drastically improves repeated Gradle builds, as
previously the broken cache of our tasks would have invalidated almost
the complete task graph. Now, almost no task has to be executed instead.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>